bots/requirements.txt
py3dtiles
discord
streamlit
kafka-python
python-dotenv
watchdog
pytest
fastapi
dpytest
twarc
fastapi[all]
